Prepping kickoff & punting game simulation improvements

+++ src/main/java/Simulation/PlayerK.java
@@ -276,12 +276,12 @@ public class PlayerK extends Player {
 
     @Override
     public int getHeismanScore() {
-        return (int) ((statsFGMade * 20 + statsXPMade*5) * (int)getFGpct()/100 + ratOvr * 10);
+        return (statsFGMade * 20 + statsXPMade*5) * (int)getFGpct()/100 + ratOvr * 10;
     }
 
     @Override
     public int getCareerScore() {
-        return (int) ((statsFGMade * 20 + statsXPMade*5) * (int)getFGpct()/100 + ratOvr * 10 + (int)(careerFGMade * 25 + careerXPMade *5) * (int)(getCareerFGpct()/100) + ratOvr*10*year);
+        return (statsFGMade * 20 + statsXPMade*5) * (int)getFGpct()/100 + ratOvr * 10 + (careerFGMade * 25 + careerXPMade *5) * (int)(getCareerFGpct()/100) + ratOvr*10*year;
     }
 
     public double getFGpct() {
